data/requirements.txt
appdirs==1.4.4
argcomplete==3.6.2
astroid==3.3.8
asttokens==2.4.1
attrs==25.1.0
beautifulsoup4==4.13.4
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.38.12
botocore==1.38.12
Brlapi==0.8.6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
crypt_r==3.13.1
cson==0.8
cssselect==1.2.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
dill==0.3.9
distro==1.9.0
dnf==4.23.0
dnspython==2.7.0
docutils==0.21.2
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.15.4
humanize==4.12.0
idna==3.10
initial-setup==0.3.101
inkex==1.4.0
iso639==0.1.4
isort==6.0.1
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langtable==0.0.68
libcomps==0.1.21
libdnf==0.74.0
libdnf5==5.2.13.1
LinkChecker==10.5.0
louis==3.28.0
lxml==5.3.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.15.0
mypy_extensions==1.0.0
nftables==0.1
numpy==2.2.5
olefile==0.47
packaging==24.2
parso==0.8.4
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
platformdirs==4.2.2
ply==3.11
productmd==1.45
psutil==6.1.1
psycopg==3.2.9
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycups==2.0.4
pyenchant==3.2.2
pygdbmi==0.11.0.0
Pygments==2.18.0
PyGObject==3.50.0
pyinotify==0.9.6
pykickstart==3.62
pylint==3.3.7
pyparsing==3.1.2
pyparted==3.13.0
PyQt5==5.15.11
PyQt5_sip==12.16.1
pyserial==3.5
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-gettext==4.0
python-linux-procfs==0.7.3
python-manatools==0.0.4
python-meh==0.52
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.14.5-build/kernel-6.14.5/linux-6.14.5-300.fc42.x86_64/tools/net/ynl
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rich==14.0.0
rpds-py==0.24.0
rpm==4.20.1
s3transfer==0.12.0
scour==0.38.2
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
Send2Trash==1.8.2
sentry-sdk==2.21.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.35-build/setroubleshoot-3.3.35/src
setuptools==74.1.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sos==4.8.2
soupsieve==2.7
speg==0.3
systemd-python==235
thonny==4.1.6
tinycss2==1.4.0
tomlkit==0.13.2
tqdm==4.67.1
trash-cli==0.24.5.26
typing_extensions==4.12.2
urllib3==2.3.0
webencodings==0.5.1
wheel==0.45.1
xkbregistry==0.3
